refactor: move updated skeleton and original skeleton to sub directories
