+++ crates/committer/src/patricia_merkle_tree.rs
@@ -1,8 +1,5 @@
 pub mod errors;
 pub mod filled_tree;
-pub mod original_skeleton_calc;
-pub mod original_skeleton_node;
 pub mod original_skeleton_tree;
 pub mod types;
-pub mod updated_skeleton_node;
 pub mod updated_skeleton_tree;
