+++ crates/committer/src/patricia_merkle_tree/original_skeleton_tree/node.rs
